(School.java) Converted diet variable to double
(DietIndicator.java)
Fixed a bug, the sum of the % was not 100 because I included empty stomaches.
Percent now expressed from 0 to 100 (it was 0 to 1)

git-svn-id: https://subversion.renater.fr/authscm/barrier.n/svn/osmose/trunk@595 45b28200-7c74-4d1a-96bc-342942c5875a

+++ src/fr/ird/osmose/School.java
@@ -102,7 +102,7 @@ public class School extends GridPoint {
     /**
      * Matrix of diets in a time step. diet[NSPECIES+NPLANKTON][NDIETSTAGES]
      */
-    public float[][] diet;
+    public double[][] diet;
     /**
      * Number of dead individuals due to fishing.
      */
@@ -188,12 +188,12 @@ public class School extends GridPoint {
         // Reset variables
         catchable = true;
         // Reset diet variables
-        diet = new float[getConfiguration().getNSpecies() + getConfiguration().getNPlankton()][];
+        diet = new double[getConfiguration().getNSpecies() + getConfiguration().getNPlankton()][];
         for (int i = 0; i < getConfiguration().getNSpecies(); i++) {
-            diet[i] = new float[nDietStage[i]];
+            diet[i] = new double[nDietStage[i]];
         }
         for (int i = 0; i < getConfiguration().getNPlankton(); i++) {
-            diet[i + getConfiguration().getNSpecies()] = new float[1];
+            diet[i + getConfiguration().getNSpecies()] = new double[1];
         }
     }
